# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/users:
    get:
      operationId: listUsers
      tags: [admin]
      summary: List all users (superuser + admin only)
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminUser"
        "401":
          description: Authentication required
        "403":
          description: Insufficient permissions
    post:
      operationId: createUser
      tags: [admin]
      summary: Create a new user (superuser + admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserInput"
      responses:
        "201":
          description: Created user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "400":
          description: Validation error
        "401":
          description: Authentication required
        "403":
          description: Insufficient permissions
        "409":
          description: Email already registered
  /admin/users/{id}:
    patch:
      operationId: updateUser
      tags: [admin]
      summary: Update a user (role, isApproved, name, password reset)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "404":
          description: User not found
    delete:
      operationId: deleteUser
      tags: [admin]
      summary: Delete a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: User not found
...
components:
  schemas:
...
    AdminUser:
      type: object
      required: [id, email, name, role, isApproved, isBlocked, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
        isApproved:
          type: boolean
        isBlocked:
          type: boolean
        linkedEntityId:
          type: string
          nullable: true
        hasPassword:
          type: boolean
        hasGoogleId:
          type: boolean
        createdAt:
          type: string
          format: date-time
    AdminUserUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        role:
          type: string
        isApproved:
          type: boolean
        isBlocked:
          type: boolean
        linkedEntityId:
          type: string
          nullable: true
        newPassword:
          type: string
          minLength: 6
          nullable: true
...
    CreateUserInput:
      type: object
      required: [email, name, role, password, isApproved]
      properties:
        email:
          type: string
          format: email
        name:
          type: string
          minLength: 1
        role:
          type: string
        password:
          type: string
          minLength: 6
        isApproved:
          type: boolean
        linkedEntityId:
          type: string
          nullable: true
...