# lib/api-zod/src/generated/types/adminUser.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AdminUser {
  id: number;
  email: string;
  name: string;
  role: string;
  isApproved: boolean;
  isBlocked: boolean;
  linkedEntityId?: string | null;
  hasPassword?: boolean;
  hasGoogleId?: boolean;
  createdAt: Date;
}