# lib/api-zod/src/generated/types/adminUserUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AdminUserUpdate {
  /** @minLength 1 */
  name?: string;
  role?: string;
  isApproved?: boolean;
  isBlocked?: boolean;
  linkedEntityId?: string | null;
  /** @minLength 6 */
  newPassword?: string | null;
}